7월10일 커밋

+++ back-end/src/main/java/shop/watchawiki/api/movies/controller/MoviesController.java
@@ -27,7 +27,7 @@ public class MoviesController {
 
     @GetMapping("/count")
     public int count(){
-        return moviesService.count();
+        return (int)moviesService.count();
     }
 
     @DeleteMapping("/{movieId}")
